scala - 2023 day 11

+++ scala/utils/src/main/scala/utils/Puzzle.scala
@@ -2,13 +2,13 @@ package utils
 
 trait Puzzle {
 
-  def solve(resource: String): (Int, Int) = {
+  def solve(resource: String): (Long, Long) = {
     val input = Reader(resource).read()
     (part1(input), part2(input))
   }
 
-  def part1(input: List[String]): Int
+  def part1(input: List[String]): Long
 
-  def part2(input: List[String]): Int
+  def part2(input: List[String]): Long
 
 }
